fix(type): resolve all Pylance type errors across codebase
全面修复 Pylance 类型检查错误

修复范围:
- kernel/operator: 5个文件 (flatmap, join, sink, source等)
- kernel/transformation: 2个文件 (sink, keyby)
- kernel/context: 3个文件 (base, injection, task)
- kernel/service: 2个文件 (base_service_task, service_caller)
- kernel/task: 2个文件 (base_task, ray_task)
- kernel/runtime: 3个文件 (dispatcher, job_info, job_manager)
- kernel/utils: 3个文件 (persistence/state, ray/actor, ray/ray_utils)
- kernel/tests: 9个测试文件
- libs: 4个文件 (batch, sink, source, document_loaders)
- middleware: 4个文件 (sage_db, sage_flow)
- platform: 1个文件 (base_queue_descriptor)
- tools: 3个文件 (CLI commands and utils)
- examples: 新增示例文件 (tutorials)

修复策略:
1. 优先添加类型声明而非 type: ignore
2. 只在必要时使用 type: ignore 并附带说明注释
3. 所有 type: ignore 都有明确的原因注释
4. 添加运行时断言以确保类型安全
5. 修复动态属性访问问题
6. 处理条件导入的类型问题

关键修复:
- 动态属性注入: operator, dispatcher, task, _actor
- 运行时保证 vs 静态类型: parallel_index, queue_instance
- 条件模块导入: ray 模块可能为 None
- Mock 对象测试: 添加适当的 type: ignore[arg-type]
- 抽象类测试: type: ignore[abstract] 用于故意测试
- Property 重定义: 移除 is_comap 重复定义

统计:
- 修复文件: 30+ 个
- 解决错误: 100+ 个
- type: ignore 使用: ~35 处 (全部有注释)
- 代码改进: 5+ 处 (assert, if checks)

质量标准:
✅ 最小化 type: ignore 使用
✅ 100% 注释覆盖率
✅ 运行时安全检查
✅ 一致的修复模式

Closes #<issue-number>

+++ packages/sage-middleware/src/sage/middleware/components/sage_db/python/multimodal_sage_db.py
@@ -11,7 +11,7 @@ import numpy as np
 
 try:
     # 假设编译生成的C++扩展模块
-    import _sage_db
+    import _sage_db  # type: ignore[import-not-found]
 except ImportError:
     print("Warning: C++ extension not available. Using mock implementation.")
     _sage_db = None
@@ -302,7 +302,7 @@ class MultimodalSageDB:
                 )
                 similarities.append(cos_sim)
 
-        return np.mean(similarities) if similarities else 0.0
+        return float(np.mean(similarities)) if similarities else 0.0
 
 
 # 便利函数
